+++ lib/screens/account_tab/account_tab.dart
@@ -1,9 +1,9 @@
-import 'package:anonaddy/constants.dart';
 import 'package:anonaddy/screens/account_tab/domains.dart';
 import 'package:anonaddy/screens/account_tab/recipients.dart';
+import 'package:anonaddy/shared_components/constants/material_constants.dart';
+import 'package:anonaddy/shared_components/loading_indicator.dart';
+import 'package:anonaddy/shared_components/lottie_widget.dart';
 import 'package:anonaddy/state_management/providers/global_providers.dart';
-import 'package:anonaddy/widgets/loading_indicator.dart';
-import 'package:anonaddy/widgets/lottie_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
