+++ lib/screens/account_tab/additional_username.dart
@@ -1,10 +1,10 @@
-import 'package:anonaddy/constants.dart';
 import 'package:anonaddy/screens/account_tab/username_detailed_screen.dart';
+import 'package:anonaddy/shared_components/constants/toast_messages.dart';
+import 'package:anonaddy/shared_components/custom_page_route.dart';
+import 'package:anonaddy/shared_components/lottie_widget.dart';
+import 'package:anonaddy/shared_components/shimmer_effects/recipients_shimmer_loading.dart';
 import 'package:anonaddy/state_management/providers/class_providers.dart';
 import 'package:anonaddy/state_management/providers/global_providers.dart';
-import 'package:anonaddy/widgets/custom_page_route.dart';
-import 'package:anonaddy/widgets/lottie_widget.dart';
-import 'package:anonaddy/widgets/shimmer_effects/recipients_shimmer_loading.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/painting.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
