+++ lib/screens/account_tab/main_account.dart
@@ -1,14 +1,16 @@
 import 'package:anonaddy/models/user/user_model.dart';
 import 'package:anonaddy/screens/account_tab/add_new_recipient.dart';
+import 'package:anonaddy/shared_components/account_list_tile.dart';
+import 'package:anonaddy/shared_components/constants/official_anonaddy_strings.dart';
+import 'package:anonaddy/shared_components/constants/toast_messages.dart';
+import 'package:anonaddy/shared_components/constants/url_strings.dart';
 import 'package:anonaddy/state_management/providers/class_providers.dart';
 import 'package:anonaddy/state_management/providers/global_providers.dart';
 import 'package:anonaddy/utilities/niche_method.dart';
-import 'package:anonaddy/widgets/account_list_tile.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:url_launcher/url_launcher.dart';
 
-import '../../constants.dart';
 import 'add_new_username.dart';
 
 class MainAccount extends StatelessWidget {
@@ -90,7 +92,7 @@ class MainAccount extends StatelessWidget {
 
   Future updateDefaultAliasFormatDomain(
       BuildContext context, Function showToast) async {
-    await launch(kDefaultAliasURL).catchError((error, stackTrace) {
+    await launch(kAnonAddySettingsURL).catchError((error, stackTrace) {
       throw showToast(error.toString());
     });
   }
