+++ lib/screens/account_tab/recipients.dart
@@ -1,7 +1,7 @@
+import 'package:anonaddy/shared_components/lottie_widget.dart';
+import 'package:anonaddy/shared_components/recipient_list_tile.dart';
+import 'package:anonaddy/shared_components/shimmer_effects/recipients_shimmer_loading.dart';
 import 'package:anonaddy/state_management/providers/global_providers.dart';
-import 'package:anonaddy/widgets/lottie_widget.dart';
-import 'package:anonaddy/widgets/recipient_list_tile.dart';
-import 'package:anonaddy/widgets/shimmer_effects/recipients_shimmer_loading.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
