+++ lib/services/alias/alias_service.dart
@@ -5,10 +5,10 @@ import 'dart:io';
 import 'package:anonaddy/models/alias/alias_data_model.dart';
 import 'package:anonaddy/models/alias/alias_model.dart';
 import 'package:anonaddy/services/data_storage/offline_data_storage.dart';
+import 'package:anonaddy/shared_components/constants/url_strings.dart';
 import 'package:anonaddy/utilities/api_message_handler.dart';
 import 'package:http/http.dart' as http;
 
-import '../../constants.dart';
 import '../access_token/access_token_service.dart';
 
 class AliasService {
