+++ lib/services/domains/domains_service.dart
@@ -4,10 +4,9 @@ import 'dart:io';
 import 'package:anonaddy/models/domain/domain_model.dart';
 import 'package:anonaddy/services/access_token/access_token_service.dart';
 import 'package:anonaddy/services/data_storage/offline_data_storage.dart';
+import 'package:anonaddy/shared_components/constants/url_strings.dart';
 import 'package:anonaddy/utilities/api_message_handler.dart';
 import 'package:http/http.dart' as http;
-
-import '../../constants.dart';
 
 class DomainsService {
   final _accessTokenService = AccessTokenService();
