+++ lib/services/username/username_service.dart
@@ -5,10 +5,9 @@ import 'package:anonaddy/models/username/username_data_model.dart';
 import 'package:anonaddy/models/username/username_model.dart';
 import 'package:anonaddy/services/access_token/access_token_service.dart';
 import 'package:anonaddy/services/data_storage/offline_data_storage.dart';
+import 'package:anonaddy/shared_components/constants/url_strings.dart';
 import 'package:anonaddy/utilities/api_message_handler.dart';
 import 'package:http/http.dart' as http;
-
-import '../../constants.dart';
 
 class UsernameService {
   final _accessTokenService = AccessTokenService();
