+++ lib/shared_components/alias_list_tile.dart
@@ -1,11 +1,12 @@
 import 'package:anonaddy/models/alias/alias_data_model.dart';
 import 'package:anonaddy/screens/alias_tab/alias_detailed_screen.dart';
 import 'package:anonaddy/state_management/providers/class_providers.dart';
-import 'package:anonaddy/widgets/alias_list_tile_leading.dart';
-import 'package:anonaddy/widgets/custom_page_route.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+import 'alias_list_tile_leading.dart';
+import 'custom_page_route.dart';
 
 class AliasListTile extends ConsumerWidget {
   const AliasListTile({Key key, this.aliasData}) : super(key: key);
