+++ lib/shared_components/constants/material_constants.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+/// Colors
+const kBlueNavyColor = Color(0xFF19216C);
+const kAccentColor = Color(0xFF62F4EB);
+
+/// UI Decoration
+const kTextFormFieldDecoration = InputDecoration(
+  border: OutlineInputBorder(),
+  focusedBorder:
+      OutlineInputBorder(borderSide: BorderSide(color: kAccentColor)),
+);
