+++ lib/shared_components/constants/official_anonaddy_strings.dart
@@ -1,0 +1,32 @@
+/// Official AnonAddy Messages
+const kDeletedAliasText =
+    'Deleted aliases reject all emails sent to them. However, they can be restored to start receiving emails again.';
+const kDeleteAliasConfirmation =
+    'Are you sure you want to delete this alias? You can restore this alias if you later change your mind. Once deleted, this alias will reject any emails sent to it.';
+const kCreateNewAliasText =
+    'Other aliases e.g. alias@khalidwar.anonaddy.com or .me can also be created automatically when they receive their first email.';
+const kRestoreAliasText =
+    'Are you sure you want to restore this alias? Once restored, this alias will be able to receive emails again.';
+const kRemoveRecipientPublicKey = 'Remove recipient public key';
+const kRemoveRecipientPublicKeyBody =
+    'Are you sure you want to remove the public key for this recipient?\nIt will also be removed from any other recipients using the same key.';
+const kDeleteRecipientDialogText =
+    'Are you sure you want to delete this recipient?';
+const kAddRecipientText =
+    'Enter the individual email of the new recipient you\'d like to add.\n\nYou will receive an email with a verification link that will expire in one hour, you can click "Resend email" to get a new one.';
+const kUpdateAliasRecipients =
+    'Select the recipients for this alias. You can choose multiple recipients.\n\nLeave it empty if you would like to use the default recipient.';
+const kAddNewUsernameText =
+    'Please choose additional usernames carefully as you can only add a maximum of three. You cannot login with these usernames, only the one you originally signed up with.';
+const kDeleteUsername =
+    'Are you sure you want to delete this username? This will also delete all aliases associated with this username.\n\nYou will no longer be able to receive any emails at this username subdomain.\n\nThis will still count towards your additional username limit even once deleted.';
+const kReachedUsernameLimit = 'You have reached your additional username limit';
+const kReachedRecipientLimit = 'You have reached your recipient limit';
+const kUpdateUsernameDefaultRecipient =
+    'Select the default recipient for this username. This overrides the default recipient in your account settings.\n\nLeave it empty if you would like to use the default recipient in your account settings.';
+
+/// Official AnonAddy Code Strings
+const kUUID = 'uuid';
+const kRandomChars = 'random_characters';
+const kCustom = 'custom';
+const kRandomWords = 'random_words';
