+++ lib/shared_components/constants/toast_messages.dart
@@ -1,0 +1,16 @@
+/// Toast Messages
+const kCopiedToClipboard = 'Copied to Clipboard';
+const kAliasRestoredSuccessfully = 'Alias Restored Successfully';
+const kFailedToRestoreAlias = 'Failed to Restore Alias';
+const kAliasDeletedSuccessfully = 'Alias Deleted Successfully';
+const kFailedToDeleteAlias = 'Failed to Delete Alias';
+const kEncryptionEnabled = 'Encryption enabled';
+const kFailedToEnableEncryption = 'Failed to Enable Encryption';
+const kEncryptionDisabled = 'Encryption disabled';
+const kFailedToDisableEncryption = 'Failed to disable encryption';
+const kGPGKeyDeletedSuccessfully = 'GPG Key Deleted Successfully';
+const kFailedToDeleteGPGKey = 'Failed to Delete GPG Key';
+const kGPGKeyAddedSuccessfully = 'GPG Key Added Successfully';
+const kFailedToAddGPGKey = 'Failed to Add GPG Key';
+const kOnlyAvailableToPaid = 'Only available to paid users';
+const kComingSoon = 'Coming Soon';
