+++ lib/shared_components/constants/ui_strings.dart
@@ -1,0 +1,24 @@
+/// UI Strings
+const kSearchHintText = 'Search aliases by email or description';
+const kLoadingText = 'Loading';
+const kDescriptionInputText = 'Description (optional)';
+const kEnterLocalPart = 'Enter Local Part (no space or @)';
+const kNoInternetConnection =
+    'No Internet Connection.\nMake sure you\'re online.';
+const kEditDescSuccessful = 'Description Updated Successfully';
+const kEditDescFailed = 'Failed to update description';
+const kGetAccessToken =
+    'To access your AnonAddy account, you have to obtain an API Access Token.'
+    '\n\nLogin to your AnonAddy account, head to settings, and navigate to API section.'
+    '\n\nGenerate a new access token.'
+    '\n\nAccess Token is a long string of alphabets, numbers, and special characters.'
+    '\n\nPaste it here as is!';
+const kPublicGPGKeyHintText =
+    'Begins with \'-----BEGIN PGP PUBLIC KEY BLOCK-----\'';
+const kEnterPublicKeyData =
+    'Enter your PUBLIC key data in the text area below. Make sure to remove Comment: and Version:';
+const kSignOutAlertDialog = 'Are you sure you want to sign out?';
+const kUnverifiedRecipient =
+    'Unverified recipient emails can NOT be set as default recipient for aliases.';
+const kUpdateAliasRecipientFailed = 'Failed to update description';
+const kUpdateAliasRecipientSuccessful = 'Successfully updated alias recipients';
