+++ lib/shared_components/constants/url_strings.dart
@@ -1,0 +1,26 @@
+/// AnonAddy URLs
+const kBaseURL = 'https://app.anonaddy.com/api/v1';
+const kAccountDetailsURL = 'account-details';
+const kActiveAliasURL = 'active-aliases';
+const kAliasURL = 'alias';
+const kAliasesURL = 'aliases';
+const kEncryptedRecipient = 'encrypted-recipients';
+const kRecipientsURL = 'recipients';
+const kRecipientKeys = 'recipient-keys';
+const kActiveUsernamesURL = 'active-usernames';
+const kUsernamesURL = 'usernames';
+const kCatchAllUsernameURL = 'catch-all-usernames';
+const kDefaultRecipientURL = 'default-recipient';
+const kDomainsURL = 'domains';
+const kDomainsOptionsURL = 'domain-options';
+const kAnonAddySettingsURL = 'https://app.anonaddy.com/settings';
+
+/// AddyManager URLs
+const kAddyManagerRepoURL = 'https://github.com/KhalidWar/anonaddy';
+const kAddyManagerIssuesURL = 'https://github.com/KhalidWar/anonaddy/issues';
+const kAddyManagerLicenseURL =
+    'https://github.com/KhalidWar/anonaddy/blob/master/LICENSE';
+
+/// Persons
+const kKhalidWarGithubURL = 'https://github.com/KhalidWar';
+const kWillBrowningGithubURL = 'https://github.com/willbrowningme';
