+++ lib/shared_components/lottie_widget.dart
