+++ lib/shared_components/no_internet_alert.dart
