+++ lib/shared_components/shimmer_effects/alias_shimmer_loading.dart
@@ -1,8 +1,8 @@
-import 'package:anonaddy/widgets/alias_tab_pie_chart.dart';
+import 'package:anonaddy/shared_components/constants/material_constants.dart';
 import 'package:flutter/material.dart';
 import 'package:shimmer/shimmer.dart';
 
-import '../../constants.dart';
+import '../alias_tab_pie_chart.dart';
 
 class AliasShimmerLoading extends StatelessWidget {
   @override
